Avoid file cache collision on the same filename
Using the DropboxDownloader it's possible to have a collision when
looking for a file into the cache when filename have the same value.
To avoid it, we changed the implementation of GetFilename() when a
downloadRecord has existing arguments and use their values to compute a
hashedArgs used to build the filepath in the cache.